organize model package a wee bit
