+++ src/main/java/tla/backend/es/model/LemmaEntity.java
@@ -19,6 +19,9 @@ import lombok.NoArgsConstructor;
 import lombok.Singular;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
+import tla.backend.es.model.meta.TLAEntity;
+import tla.backend.es.model.parts.LemmaWord;
+import tla.backend.es.model.parts.Translations;
 import tla.domain.dto.LemmaDto;
 import tla.domain.model.meta.BTSeClass;
 import tla.domain.model.meta.TLADTO;
