+++ src/main/java/tla/backend/es/model/meta/TLAEntity.java
@@ -1,4 +1,4 @@
-package tla.backend.es.model;
+package tla.backend.es.model.meta;
 
 import java.util.Collections;
 import java.util.List;
