+++ src/main/java/tla/backend/es/model/parts/EditorInfo.java
@@ -1,4 +1,4 @@
-package tla.backend.es.model;
+package tla.backend.es.model.parts;
 
 import java.util.Date;
 import java.util.List;
